refactor ArchipelagoManager into 4 additional classes (Options, ItemHandling, LocationHandling, and LocationAccessibility)
