+++ GatorRando/patches/ItemSearchObjectsPatch.cs
@@ -1,4 +1,5 @@
 using System;
+using GatorRando.Archipelago;
 using GatorRando.UIMods;
 using HarmonyLib;
 
@@ -13,8 +14,8 @@ static class ItemSearchObjectsPatch
     {
         __result = SettingsMods.GetCheckfinderBehavior() switch
         {
-            SettingsMods.CheckfinderBehavior.Logic => Array.FindAll(__result, ArchipelagoManager.IsLocationAccessible),
-            SettingsMods.CheckfinderBehavior.ChecksOnly => Array.FindAll(__result, ArchipelagoManager.IsLocationACheck),
+            SettingsMods.CheckfinderBehavior.Logic => Array.FindAll(__result, LocationAccessibilty.IsLocationAccessible),
+            SettingsMods.CheckfinderBehavior.ChecksOnly => Array.FindAll(__result, LocationAccessibilty.IsLocationACheck),
             SettingsMods.CheckfinderBehavior.Original => __result,
             _ => throw new Exception("Invalid enum value for CheckfinderBehavior"),
         };
