+++ GatorRando/questMods/QuestItems.cs
@@ -34,6 +34,9 @@ static class QuestItems
             QuestItemObjects.Add(Util.GenerateItemObject("Thrown_Pencil", pencil.sprite));
             QuestItemObjects.Add(Util.GenerateItemObject("Thrown_Pencil_2", pencil.sprite));
             QuestItemObjects.Add(Util.GenerateItemObject("Thrown_Pencil_3", pencil.sprite));
+
+            //TODO: Retrieve/Store sprites in a way that does not result in them unloading
+            //TODO: Populate items for pot, chest, race items
         }
     }
 }
